YCMD Completions Integration
This patch gets the server working end to end with the swift completion
client in YCMD ( my fork atleast ).

Add a shutdown endpoint so we can restart the sever easily.

Add the ability to pass in options from the client including HMAC secret
and logging level.

Add a bootstrap script to easily get started.

Additionally it includes other cleanups:

- Use the correct JSON headers
- Remove unused code and some naming fixes
- Remove prototype code cleanup names
- Cleanup build defs
- Remove old server code