Refactor logging
Abstract out raw use of std::out. It adds the notion of a log level,
which allows switch on and off between diffent levels. Also it adds
logging channels. A logging channel is used to namespace messages in the
log. In the future, this may select to custom streams or files.

I adopted the new logging system in the entire app.

Additionally:
- turn off crazy server logs.
- turn off logging in SwiftCompleter.
- make clang format explicit.