Refactor SourceKitService and Propagate `log`
This patch refactors sourcekitd into a service class. Its a bit easier
work with and to reason about it now I think.

I made logging uniform across the entire SourceKitService: logsl happens
within the service and at the same time in a given request. Dumps of
responses are logged as LogLevelExtreme.

Additionally, I added the ability to use the log level set by the
command line. And modeled the logic to use an equality. It seems more
representative of the relation between these types.