Spartan API Integration Tests
Write some integration tests for the HTTP API.

This tests the system end to end and makes sure that responses are
returned. It hits the API with some basic HTTP requests.

While spartan, this adds some assurance it is still working end to end.

It uses "assert" as a basic mechanism to check state.

Additional asserts happen in code via boost::variant and a lightweight
"Result" type. Incorrect result access will cause an exception, which is
exactly what the tests do: always access the result they are expecting.

The testing system keeps C++ as the default language, even though it
required some extra typing. Some of the helper functions are reusable
and can go in other tests in the future.

I deleted the "req.sh" program since this is a better way to make
requests.