Decouple from OSX, Xcode, Brew
This patch decouples the build from OSX, Xcode, and Brew.

This represents the theoretical state of dependencies: there's no
dependency on OSX, Xcode or Brew!

These build parameters are expressed as fallback values: they are only
used if the user has not provided their own parameters.

Refactor CMakeLists in a way that is more configurable
- allow the user to provide a boost installation
- allow the user to provide a sourcekit installation

Now, all source code related operations of the build system are governed
by CMake: bootstrapping just calls configure and build.

The fact that it defaults to homebrew for boost isn't really a
requirement, but it is what it uses for now. This is a marginal default
and I added a FIXME, to improve it in the future.

Note: As of writing, my development environments meet these requirements
and it seems like sensible default for many potential users. As much as
I don't enjoy Xcode, this is pragmatic for reasons described in the
README, and CMakeLists, and Swift YCMD RFC.