Remove operator completion hack
We were originally returning completions for operator by hacking the
completion offset. This may have been intentional at one point ( or an
off by 1); after using this a lot and @puremournings feedback, it's
removed.

Operators are no longer returned, which is more reasonable behavior: if
a codebase is using operator overloading to the point at which
completion for operators is useful, they have other issues that are out
of scope of this work.