Remove brew install boost
This patch improves the clean build performance and predictability of
the build by removing the previous `$brew pray and spray` build tactic
for boost.

Check in boost and build it with boost's build system via cmake. This is
easier and uses boost's existing build abilities. Use boost-cp to
intelligently fetch minimum set of dependent files. Then, further
processing strips it down even more to minimize the repo footprint.

Included is the script I used to achieve this with BCP, which needs some
work to update again it again but is highly useful.