WIP: Move flag prep logic to server
In hindsight, it was a shitty decision to put flag prep in the client (
Probably a whisky driven design decision ) Revert this mistake.

The CompilationDatabase should be an input to Swift services and tools (
like how this works in clang ).

Now the client simply prepares a CompilationDatabase, and sends us
commands from that.

We need to process flags for diagnostics and it is more reasonable to
have all of this flag logic happen in 1 place, and have clear
definitions of what the server accepts as input for these flags.

resolves #6

There should be a PR to YCMD as well ( or just merge into
https://github.com/Valloric/ycmd/pull/487 )

TODO:
Additionally consider making the flags accept a string instead of an
array. It is insane to convert the string to an array on the client.