Fix common completion and diagnostic crashes
This patch fix a few empty variant issues to address the known crashes
in completions and diagnostics.

The server wasn't correctly handling odd behavior from SourceKit, which
would take it down.

- Sema notifications would trigger a crash
- Empty or NULL responses would trigger a crash in both the server and client.
- Add a mutex around `std::cout` and `std::cerr`.

For now, it write empty responses for completions and diagnostics to
keep the client running. Treating errors as empty responses isn't ideal:
The old ycmd client freeze the completion at the identifier value. This
should be addressing this in the new vim client. The should eventually
(potentially) be propagated to the JSON API, and handled in the new
client.

Reproducibility:

I was able to see these errors on 8.3.2 with the old client and test
data: `ycmd/tests/swift/testdata/iOS/Basic/Basic/`. ( I think that fake
comp db is weird or something ).  Similar issues are also happing in
SwiftVimTestHost